refactor(client): consolidate formatGoldValue into currency utils
Move OSRS-style gold formatting to the canonical location in
game/systems/currency/currencyUtils.ts. Remove duplicate definitions
from DuelPanel/utils.ts and TradePanel/utils.ts. Update all consumers
(StakesScreen, ConfirmScreen, DuelResultModal, StakeGrid, TradePanel)
to import from the central currency system.
