+++ packages/client/src/game/panels/DuelPanel/utils.ts
@@ -18,20 +18,6 @@ export function formatQuantity(quantity: number): string {
 }
 
 /**
- * Format gold value with K/M/B suffixes
- */
-export function formatGoldValue(value: number): string {
-  if (value >= 1_000_000_000) {
-    return `${(value / 1_000_000_000).toFixed(1)}B`;
-  } else if (value >= 1_000_000) {
-    return `${(value / 1_000_000).toFixed(1)}M`;
-  } else if (value >= 1_000) {
-    return `${(value / 1_000).toFixed(1)}K`;
-  }
-  return value.toString();
-}
-
-/**
  * Calculate total value of staked items
  */
 export function calculateTotalValue(
